+++ capi/CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   - Added `type_size` member to input types and `BLAZE_INPUT` macro for
     initialization
   - Reserved trailing padding bytes to ensure zero initialization
+  - Reserved space for future extension in output types
 - Added `blaze_normalizer_new_opts` function and `blaze_normalizer_opts`
   type
 - Renamed various symbolization functions to closer reflect Rust